[MOVE] <introduce module> set file structure
