+++ src/introduce/club/entity/club.repository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from "typeorm";
-import { ClubList } from "../club.dto";
+import { ClubList } from "../dto/club.dto";
 import { Club } from "./club.entity";
 
 @EntityRepository(Club)
