+++ src/introduce/club/user/user.entity.ts
@@ -1,4 +1,4 @@
-import { Column, Entity, OneToMany, PrimaryColumn, PrimaryGeneratedColumn } from "typeorm";
+import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from "typeorm";
 import { ClubMember } from "../entity/club.member.entity";
 
 @Entity("user")
